Added UserProfile component

+++ app/Http/Controllers/Api/UserAvatarController.php
@@ -69,7 +69,7 @@ class UserAvatarController extends Controller
     {
         $oldImage = public_path(auth()->user()->meta->avatar_path);
 
-        if (File::exists($oldImage)) {
+        if (File::exists($oldImage) && File::isFile($oldImage)) {
             unlink($oldImage);
         }
     }
